+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-BVqzx1Ax.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-DAgyBYNl.js";import{_ as a}from"./mermaid.core-D04Syzk5.js";import"./chunk-5VM5RSS4-BtTRVOFR.js";import"./chunk-XXDRQBXY-Atx7gZfX.js";import"./chunk-POPQ4Y6H-DGRW6V_C.js";import"./chunk-F27PBJKO-huARZnCy.js";import"./channel-DLBacKT2.js";import"./index-CZcZxAwO.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};