+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CSm4FfiW.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-IBnkm4Kc.js";import{_ as a}from"./mermaid.core-BxaWNiiY.js";import"./chunk-5VM5RSS4--DdSXdap.js";import"./chunk-XXDRQBXY-By0mwZju.js";import"./chunk-POPQ4Y6H-Dns-dMlF.js";import"./chunk-F27PBJKO-CXsP48Nt.js";import"./channel-BREoxQWm.js";import"./index-CYbD4QzK.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};